ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pymupdf==1.25.1
python-docx==1.1.2
scikit-learn==1.6.0
numpy==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.1